Added data to benchmark StringIntern
