+++ Homework1/Benchmarks/Program.cs
@@ -1,37 +1,5 @@
-﻿using System.Text;
-using BenchmarkDotNet.Attributes;
-using BenchmarkDotNet.Order;
-using BenchmarkDotNet.Running;
+﻿using BenchmarkDotNet.Running;
+using Fuse8_ByteMinds.SummerSchool.Benchmarks;
 
-BenchmarkRunner.Run<StringInternBenchmark>();
-
-[MemoryDiagnoser(displayGenColumns: true)]
-[Orderer(SummaryOrderPolicy.FastestToSlowest)]
-[RankColumn]
-public class StringInternBenchmark
-{
-    private readonly List<string> _words = new();
-    public StringInternBenchmark()
-    {
-       foreach (var word in File.ReadLines(@".\SpellingDictionaries\ru_RU.dic"))
-           _words.Add(string.Intern(word));
-    }
-
-    [Benchmark(Baseline = true)]
-    [ArgumentsSource(nameof(SampleData))]
-    public bool WordIsExists(string word)
-        => _words.Any(item => word.Equals(item, StringComparison.Ordinal));
-
-    [Benchmark]
-    [ArgumentsSource(nameof(SampleData))]
-    public bool WordIsExistsIntern(string word)
-    {
-        var internedWord = string.Intern(word);
-        return _words.Any(item => ReferenceEquals(internedWord, item));
-    }
-
-    public IEnumerable<string> SampleData()
-    {
-        yield return new StringBuilder().ToString();
-    }
-}
+//BenchmarkRunner.Run<StringInternBenchmark>();
+BenchmarkRunner.Run<AccountProcessorBenchmark>();
